refactor clicksystem's singleton implementation to use ECS features
removes the native array of clickDatas (which was informally acting as a component singleton as state on the system). using the 'SetSingleton' api, we're treating the ClickSystem as stateless now, which is a good thing.

to retrieve the click data, instead of getting a reference to the ClickSystem, and reaching into it's ClickData member - we're just querying for the ClickDataEntity (our new singleton) and retriveing the ClickData component from that entity.

this will make the waypoint system easier to implement in the future!